Ajustes varios

+++ WebAsada/Controllers/PersonsByEstateController.cs
@@ -1,13 +1,12 @@
 ﻿using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Routing;
-using System;
 using System.Threading.Tasks;
+using WebAsada.BaseObjects;
 using WebAsada.Repository;
 using WebAsada.ViewModels;
 
 namespace WebAsada.Controllers
 {
-    public class PersonsByEstateController : Controller
+    public class PersonsByEstateController : BaseController
     {  
         private readonly PersonsByEstateRepository _personsByEstateRepository;
         private readonly PersonRepository _personRepository;
@@ -27,15 +26,20 @@ namespace WebAsada.Controllers
             var person = await _personRepository.GetById(personItemVM.Id);
             var estate = await _estateRepository.GetById(id.Value);
 
-            await _personsByEstateRepository.Save(person, estate);
-               
-            return Ok(new {id.Value});
+            var result = await _personsByEstateRepository.Save(person, estate);
+
+            return (result.IsSuccess) ? Ok(new { id.Value }) : ErrorContent(result.Error);
+            
         }
 
 
         public async Task<IActionResult> DeleteOwner(int? estateId, int? personId)
         {
-            if (!personId.HasValue) return NotFound(); 
+            if (!personId.HasValue) return NotFound();
+
+
+            if (await _personsByEstateRepository.VerifyIfContainPersonRelated(estateId.Value) == 1 ) return ErrorContent("No es permitido eliminar todos los dueños de una finca");
+
             var person = await _personRepository.GetById(personId.Value);
             var estate = await _estateRepository.GetById(estateId.Value);
             await _personsByEstateRepository.Delete(person, estate);
